Update index.html
Linked my languages at the top of the page, to the evaluation with each language. So that when one of the languages mentioned on the website is clicked on, it takes the user to the respective location in which a brief evaluation is located with a detailed description of my experience.